+++ src/term.c
@@ -3181,15 +3181,22 @@ term_paint(void)
              // only if followed by space
              //&& iswspace(d[1].chr) && !d[1].cc_next
              //&& d[1].chr != 0x1680 && d[1].chr != 0x3000
-             && d[1].chr == ' ' && !d[1].cc_next
+             && ((d[1].chr == ' ' && !d[1].cc_next)
+             // ... or with setting EmojiPlacement=full,
+             // to avoid scrolling artefacts (#1261)
+                 || cfg.emoji_placement == EMPL_FULL
+                )
              // not at cursor position? does not work for emojis
              //&& !at_cursor_pos(i, j)
              // and significant attributes are equal
              && !((d->attr.attr ^ d[1].attr.attr) & ~IGNOVRHANG)
-             // do not overhang numbers, flag letters etc
+#ifdef exempt_emoji_components_from_overhang
+#warning drop this for #1261
+             // do not overhang numbers, flag letters etc (~#1104)
              && d->chr >= 0x80 // exclude #️*️0️..9️
              //&& !(ch >= 0x1F1E6 || ch <= 0x1F1FF) // exclude 🇦..🇿
              && !(!e.seq && emoji_bases[e.idx].ch >= 0x1F1E6 && emoji_bases[e.idx].ch <= 0x1F1FF)
+#endif
                )
             {
               d->attr.attr |= TATTR_OVERHANG;
@@ -3430,15 +3437,22 @@ term_paint(void)
              // only if followed by space
              //&& iswspace(d[1].chr) && !d[1].cc_next
              //&& d[1].chr != 0x1680 && d[1].chr != 0x3000
-             && d[1].chr == ' ' && !d[1].cc_next
+             && ((d[1].chr == ' ' && !d[1].cc_next)
+             // ... or with setting EmojiPlacement=full,
+             // to avoid scrolling artefacts (#1261)
+                 || cfg.emoji_placement == EMPL_FULL
+                )
              // not at cursor position? does not work for emojis
              //&& !at_cursor_pos(i, j)
              // and significant attributes are equal
              && !((d->attr.attr ^ d[1].attr.attr) & ~IGNOVRHANG)
-             // do not overhang numbers, flag letters etc
+#ifdef exempt_emoji_components_from_overhang
+#warning drop this for #1261
+             // do not overhang numbers, flag letters etc (~#1104)
              && d->chr >= 0x80 // exclude #️*️0️..9️
              //&& !(ch >= 0x1F1E6 || ch <= 0x1F1FF) // exclude 🇦..🇿
              && !(!e.seq && emoji_bases[e.idx].ch >= 0x1F1E6 && emoji_bases[e.idx].ch <= 0x1F1FF)
+#endif
                )
             {
               d->attr.attr |= TATTR_OVERHANG;